Extract request factory from Pinecone class
