+++ pinecone/control/pinecone.py
@@ -1,7 +1,6 @@
 import time
 import logging
-from typing import Optional, Dict, Any, Union
-from enum import Enum
+from typing import Optional, Dict, Union
 
 from .index_host_store import IndexHostStore
 from .pinecone_interface import PineconeDBControlInterface
@@ -12,27 +11,7 @@ from pinecone.core.openapi.db_control.api.manage_indexes_api import ManageIndexe
 from pinecone.openapi_support.api_client import ApiClient
 
 
-from pinecone.utils import (
-    normalize_host,
-    setup_openapi_client,
-    build_plugin_setup_client,
-    convert_enum_to_string,
-)
-from pinecone.core.openapi.db_control.models import (
-    CreateCollectionRequest,
-    CreateIndexForModelRequest,
-    CreateIndexForModelRequestEmbed,
-    CreateIndexRequest,
-    ConfigureIndexRequest,
-    ConfigureIndexRequestSpec,
-    ConfigureIndexRequestSpecPod,
-    DeletionProtection as DeletionProtectionModel,
-    IndexSpec,
-    IndexTags,
-    ServerlessSpec as ServerlessSpecModel,
-    PodSpec as PodSpecModel,
-    PodSpecMetadataConfig,
-)
+from pinecone.utils import normalize_host, setup_openapi_client, build_plugin_setup_client
 from pinecone.core.openapi.db_control import API_VERSION
 from pinecone.models import (
     ServerlessSpec,
@@ -43,7 +22,7 @@ from pinecone.models import (
     IndexEmbed,
 )
 from .langchain_import_warnings import _build_langchain_attribute_error_message
-from pinecone.utils import parse_non_empty_args, docslinks
+from pinecone.utils import docslinks
 
 from pinecone.data import _Index, _AsyncioIndex, _Inference
 from pinecone.enums import (
@@ -57,6 +36,7 @@ from pinecone.enums import (
     AzureRegion,
 )
 from .types import CreateIndexForModelEmbedTypedDict
+from .request_factory import PineconeDBControlRequestFactory
 
 from pinecone_plugin_interface import load_and_install as install_plugins
 
@@ -152,69 +132,6 @@ class Pinecone(PineconeDBControlInterface):
         except Exception as e:
             logger.error(f"Error loading plugins: {e}")
 
-    def __parse_tags(self, tags: Optional[Dict[str, str]]) -> IndexTags:
-        if tags is None:
-            return IndexTags()
-        else:
-            return IndexTags(**tags)
-
-    def __parse_deletion_protection(
-        self, deletion_protection: Union[DeletionProtection, str]
-    ) -> DeletionProtectionModel:
-        deletion_protection = convert_enum_to_string(deletion_protection)
-        if deletion_protection in ["enabled", "disabled"]:
-            return DeletionProtectionModel(deletion_protection)
-        else:
-            raise ValueError("deletion_protection must be either 'enabled' or 'disabled'")
-
-    def __parse_index_spec(self, spec: Union[Dict, ServerlessSpec, PodSpec]) -> IndexSpec:
-        if isinstance(spec, dict):
-            if "serverless" in spec:
-                index_spec = IndexSpec(serverless=ServerlessSpecModel(**spec["serverless"]))
-            elif "pod" in spec:
-                args_dict = parse_non_empty_args(
-                    [
-                        ("environment", spec["pod"].get("environment")),
-                        ("metadata_config", spec["pod"].get("metadata_config")),
-                        ("replicas", spec["pod"].get("replicas")),
-                        ("shards", spec["pod"].get("shards")),
-                        ("pods", spec["pod"].get("pods")),
-                        ("source_collection", spec["pod"].get("source_collection")),
-                    ]
-                )
-                if args_dict.get("metadata_config"):
-                    args_dict["metadata_config"] = PodSpecMetadataConfig(
-                        indexed=args_dict["metadata_config"].get("indexed", None)
-                    )
-                index_spec = IndexSpec(pod=PodSpecModel(**args_dict))
-            else:
-                raise ValueError("spec must contain either 'serverless' or 'pod' key")
-        elif isinstance(spec, ServerlessSpec):
-            index_spec = IndexSpec(
-                serverless=ServerlessSpecModel(cloud=spec.cloud, region=spec.region)
-            )
-        elif isinstance(spec, PodSpec):
-            args_dict = parse_non_empty_args(
-                [
-                    ("replicas", spec.replicas),
-                    ("shards", spec.shards),
-                    ("pods", spec.pods),
-                    ("source_collection", spec.source_collection),
-                ]
-            )
-            if spec.metadata_config:
-                args_dict["metadata_config"] = PodSpecMetadataConfig(
-                    indexed=spec.metadata_config.get("indexed", None)
-                )
-
-            index_spec = IndexSpec(
-                pod=PodSpecModel(environment=spec.environment, pod_type=spec.pod_type, **args_dict)
-            )
-        else:
-            raise TypeError("spec must be of type dict, ServerlessSpec, or PodSpec")
-
-        return index_spec
-
     def create_index(
         self,
         name: str,
@@ -226,34 +143,15 @@ class Pinecone(PineconeDBControlInterface):
         vector_type: Optional[Union[VectorType, str]] = VectorType.DENSE,
         tags: Optional[Dict[str, str]] = None,
     ) -> IndexModel:
-        if metric is not None:
-            metric = convert_enum_to_string(metric)
-        if vector_type is not None:
-            vector_type = convert_enum_to_string(vector_type)
-        if deletion_protection is not None:
-            dp = self.__parse_deletion_protection(deletion_protection)
-        else:
-            dp = None
-
-        tags_obj = self.__parse_tags(tags)
-        index_spec = self.__parse_index_spec(spec)
-
-        if vector_type == VectorType.SPARSE.value and dimension is not None:
-            raise ValueError("dimension should not be specified for sparse indexes")
-
-        args = parse_non_empty_args(
-            [
-                ("name", name),
-                ("dimension", dimension),
-                ("metric", metric),
-                ("spec", index_spec),
-                ("deletion_protection", dp),
-                ("vector_type", vector_type),
-                ("tags", tags_obj),
-            ]
+        req = PineconeDBControlRequestFactory.create_index_request(
+            name=name,
+            spec=spec,
+            dimension=dimension,
+            metric=metric,
+            deletion_protection=deletion_protection,
+            vector_type=vector_type,
+            tags=tags,
         )
-
-        req = CreateIndexRequest(**args)
         resp = self.index_api.create_index(create_index_request=req)
 
         if timeout == -1:
@@ -270,42 +168,14 @@ class Pinecone(PineconeDBControlInterface):
         deletion_protection: Optional[Union[DeletionProtection, str]] = DeletionProtection.DISABLED,
         timeout: Optional[int] = None,
     ) -> IndexModel:
-        cloud = convert_enum_to_string(cloud)
-        region = convert_enum_to_string(region)
-        if deletion_protection is not None:
-            dp = self.__parse_deletion_protection(deletion_protection)
-        else:
-            dp = None
-        tags_obj = self.__parse_tags(tags)
-
-        if isinstance(embed, IndexEmbed):
-            parsed_embed = embed.as_dict()
-        else:
-            # if dict, we need to parse enum values, if any, to string
-            # and verify required fields are present
-            required_fields = ["model", "field_map"]
-            for field in required_fields:
-                if field not in embed:
-                    raise ValueError(f"{field} is required in embed")
-            parsed_embed = {}
-            for key, value in embed.items():
-                if isinstance(value, Enum):
-                    parsed_embed[key] = convert_enum_to_string(value)
-                else:
-                    parsed_embed[key] = value
-
-        args = parse_non_empty_args(
-            [
-                ("name", name),
-                ("cloud", cloud),
-                ("region", region),
-                ("embed", CreateIndexForModelRequestEmbed(**parsed_embed)),
-                ("deletion_protection", dp),
-                ("tags", tags_obj),
-            ]
+        req = PineconeDBControlRequestFactory.create_index_for_model_request(
+            name=name,
+            cloud=cloud,
+            region=region,
+            embed=embed,
+            tags=tags,
+            deletion_protection=deletion_protection,
         )
-
-        req = CreateIndexForModelRequest(**args)
         resp = self.index_api.create_index_for_model(req)
 
         if timeout == -1:
@@ -405,68 +275,37 @@ class Pinecone(PineconeDBControlInterface):
         api_instance = self.index_api
         description = self.describe_index(name=name)
 
-        if deletion_protection is None:
-            dp = DeletionProtectionModel(description.deletion_protection)
-        elif isinstance(deletion_protection, DeletionProtection):
-            dp = DeletionProtectionModel(deletion_protection.value)
-        elif deletion_protection in ["enabled", "disabled"]:
-            dp = DeletionProtectionModel(deletion_protection)
-        else:
-            raise ValueError("deletion_protection must be either 'enabled' or 'disabled'")
-
-        fetched_tags = description.tags
-        if fetched_tags is None:
-            starting_tags = {}
-        else:
-            starting_tags = fetched_tags.to_dict()
-
-        if tags is None:
-            # Do not modify tags if none are provided
-            tags = starting_tags
-        else:
-            # Merge existing tags with new tags
-            tags = {**starting_tags, **tags}
-
-        pod_config_args: Dict[str, Any] = {}
-        if pod_type:
-            new_pod_type = pod_type.value if isinstance(pod_type, PodType) else pod_type
-            pod_config_args.update(pod_type=new_pod_type)
-        if replicas:
-            pod_config_args.update(replicas=replicas)
-
-        if pod_config_args != {}:
-            spec = ConfigureIndexRequestSpec(pod=ConfigureIndexRequestSpecPod(**pod_config_args))
-            req = ConfigureIndexRequest(deletion_protection=dp, spec=spec, tags=IndexTags(**tags))
-        else:
-            req = ConfigureIndexRequest(deletion_protection=dp, tags=IndexTags(**tags))
-
+        req = PineconeDBControlRequestFactory.configure_index_request(
+            description=description,
+            replicas=replicas,
+            pod_type=pod_type,
+            deletion_protection=deletion_protection,
+            tags=tags,
+        )
         api_instance.configure_index(name, configure_index_request=req)
 
     def create_collection(self, name: str, source: str):
-        api_instance = self.index_api
-        api_instance.create_collection(
-            create_collection_request=CreateCollectionRequest(name=name, source=source)
-        )
+        req = PineconeDBControlRequestFactory.create_collection_request(name=name, source=source)
+        self.index_api.create_collection(create_collection_request=req)
 
     def list_collections(self) -> CollectionList:
-        api_instance = self.index_api
-        response = api_instance.list_collections()
+        response = self.index_api.list_collections()
         return CollectionList(response)
 
     def delete_collection(self, name: str):
-        api_instance = self.index_api
-        api_instance.delete_collection(name)
+        self.index_api.delete_collection(name)
 
     def describe_collection(self, name: str):
-        api_instance = self.index_api
-        return api_instance.describe_collection(name).to_dict()
+        return self.index_api.describe_collection(name).to_dict()
 
     @staticmethod
     def from_texts(*args, **kwargs):
+        """@private"""
         raise AttributeError(_build_langchain_attribute_error_message("from_texts"))
 
     @staticmethod
     def from_documents(*args, **kwargs):
+        """@private"""
         raise AttributeError(_build_langchain_attribute_error_message("from_documents"))
 
     def Index(self, name: str = "", host: str = "", **kwargs):
