+++ tests/integration/control/serverless/test_sparse_index.py
@@ -70,4 +70,4 @@ class TestSparseIndexErrorCases:
 
         with pytest.raises(PineconeApiException) as e:
             client.create_index(**create_sl_index_params)
-        assert "Sparse vector indexes must use dotproduct" in str(e.value)
+        assert "Sparse vector indexes must use the metric dotproduct." in str(e.value)
